rename output to logger
